requirements.txt
# ---- Core Framework ----
fastapi==0.128.6
uvicorn==0.40.0
python-multipart==0.0.22
jinja2==3.1.6
pydantic==2.12.5

# ---- Document Parsing ----
pdfplumber==0.11.9
python-docx==1.2.0

# ---- NLP & Scoring ----
scikit-learn==1.8.0
numpy==2.4.2

# ---- HTTP Client (Ollama API) ----
httpx==0.28.1

# ---- PDF Report Export ----
reportlab==4.4.9
